Bring back basic support for datalabels for unsupported chart types
