+++ src/core/chartOptions/basicHorizontalBarchartOptions.ts
@@ -1,6 +1,6 @@
 import { Options } from 'highcharts';
 import { View } from "../types/view";
-import { commonBasicHorizontalBarChartOptions, commonBasicHorizontalBarChartYAxisOptions } from './barChartOptions';
+import { commonBasicHorizontalBarChartOptions, commonBasicHorizontalBarChartYAxisOptions, commonDatalabelsOptions } from './barChartOptions';
 
 export const basicHorizontalBarChartOptions = (view: View, locale: string): Options => {
     const highChartsOptions: Options = {
@@ -11,7 +11,12 @@ export const basicHorizontalBarChartOptions = (view: View, locale: string): Opti
             softMin: 0,
             max: 100
         },
-        legend: { enabled: false }
+        legend: { enabled: false },
+        plotOptions: {
+            series: {
+                dataLabels: [commonDatalabelsOptions(view, locale)]
+            }
+        }
     }
     return highChartsOptions;
 };
