+++ src/core/chartOptions/groupVerticalBarChartOptions.ts
@@ -2,7 +2,7 @@ import { Options } from 'highcharts';
 import { View } from "../types/view";
 import { getLegendLabelShorteningFunction } from './Utility/formatters';
 import { getTimeSeriesOptions } from './Utility/timeIntervals';
-import { commonBasicVerticalBarChartOptions } from './barChartOptions';
+import { commonBasicVerticalBarChartOptions, commonDatalabelsOptions } from './barChartOptions';
 
 export const groupVerticalBarChartOptions = (view: View, locale: string): Options => {
 
@@ -10,7 +10,8 @@ export const groupVerticalBarChartOptions = (view: View, locale: string): Option
         ...commonBasicVerticalBarChartOptions(view, locale),
         plotOptions: {
             series: {
-                ...getTimeSeriesOptions(view.visualizationSettings.timeVariableIntervals, view.visualizationSettings.timeSeriesStartingPoint)
+                ...getTimeSeriesOptions(view.visualizationSettings.timeVariableIntervals, view.visualizationSettings.timeSeriesStartingPoint),
+                dataLabels: [commonDatalabelsOptions(view, locale)]
             }
         },
         legend: {
