+++ src/core/chartOptions/pyramidChartOptions.ts
@@ -2,10 +2,13 @@ import { Options } from 'highcharts';
 import { View } from "../types/view";
 import { getFormattedUnits, getScreenReaderFormatterCallbackFunction, getToolTipFormatterFunction } from './Utility/formatters';
 import Translations from '../conversion/translations';
+import { defaultTheme } from "../highcharts/themes";
 
 export const pyramidChartOptions = (view: View, locale: string): Options => {
     const categories = view.columnNameGroups.map(cng => cng.map(n => n[locale]).join(', '));
     const maxValue = Math.max(...view.series.map(s => Math.max(...s.series.map(dataCell => dataCell.value ?? 0))));
+    const theme = defaultTheme(locale);
+    const dataValueLabelStyle = theme.tooltip?.style;
     return (
         {
             accessibility: {
@@ -70,7 +73,11 @@ export const pyramidChartOptions = (view: View, locale: string): Options => {
             credits: { text: `${Translations.source[locale]}: ${view.sources.map(s => s[locale]).join(', ')}` },
             plotOptions: {
                 series: {
-                    stacking: 'normal'
+                    stacking: 'normal',
+                    dataLabels: {
+                        enabled: view.visualizationSettings.showDataPoints,
+                        style: dataValueLabelStyle
+                    }
                 }
             },
             tooltip: {
